Refactor > use double precision float everywhere

+++ src/angles.rs
@@ -1,5 +1,5 @@
 use super::constants::*;
 
-pub fn degrees_to_radians(degrees: f32) -> f32 {
-    return degrees * PI / 180.0f32;
+pub fn degrees_to_radians(degrees: f64) -> f64 {
+    return degrees * PI / 180.0;
 }
